Rename sopel_module due to naming bug
Without this change the following error message prohibited running
troet.

> ImportError: cannot import name 'Mastodon' from partially initialized
> module 'mastodon' (most likely due to a circular import)

Renaming the sope_module from `mastodon` to `troet` fixes this.
