+++ sopel_modules/troet/__init__.py
@@ -4,10 +4,10 @@
 Currently expects to be used alongside as little other plugins/modules as possible. It (re)uses some short hand commands (e.g.: t, r, d)
 """
 
-from .mastodon import *
+from .troet import *
 
 __author__ = 'Felix Pape'
 __email__ = 'ziemlich@schlechte.info'
 __maintainer__ = 'fp'
 __version__ = '0.0.1'
-__repo__ = 'https://github.com/hackspace-marburg/sopel-mastodon'
+__repo__ = 'https://github.com/hackspace-marburg/troet'
